move dlt connector into core to use it also from federation
