+++ backend/src/config/schema.ts
@@ -5,6 +5,7 @@ import {
   COMMUNITY_URL,
   DECAY_START_TIME,
   DLT_ACTIVE,
+  DLT_CONNECTOR_URL,
   GDT_ACTIVE,
   GDT_API_URL,
   GMS_ACTIVE,
@@ -29,6 +30,7 @@ export const schema = Joi.object({
   COMMUNITY_SUPPORT_MAIL,
   DECAY_START_TIME,
   DLT_ACTIVE,
+  DLT_CONNECTOR_URL,
   GDT_API_URL,
   GDT_ACTIVE,
   GMS_ACTIVE,
@@ -69,12 +71,6 @@ export const schema = Joi.object({
     )
     .default('http://0.0.0.0/redeem/CL-')
     .required(),
-
-  DLT_CONNECTOR_URL: Joi.string()
-    .uri({ scheme: ['http', 'https'] })
-    .default('http://localhost:6010')
-    .when('DLT_ACTIVE', { is: true, then: Joi.required() })
-    .description('The URL for DLT connector'),
 
   EMAIL_LINK_VERIFICATION: Joi.string()
     .uri({ scheme: ['http', 'https'] })
