+++ core/src/apis/dltConnector/DltConnectorClient.test.ts
@@ -1,4 +1,4 @@
-import { CONFIG } from '@/config'
+import { CONFIG } from '../../config'
 
 import { DltConnectorClient } from './DltConnectorClient'
