+++ core/src/apis/dltConnector/index.ts
@@ -11,8 +11,8 @@ import {
 } from 'database'
 import { getLogger } from 'log4js'
 import { IRestResponse } from 'typed-rest-client'
-import { CONFIG } from '@/config'
-import { LOG4JS_BASE_CATEGORY_NAME } from '@/config/const'
+import { CONFIG } from '../../config'
+import { LOG4JS_BASE_CATEGORY_NAME } from '../../config/const'
 import { DltConnectorClient } from './DltConnectorClient'
 import { DltTransactionType } from './enum/DltTransactionType'
 import { TransactionDraft } from './model/TransactionDraft'
