+++ core/src/apis/dltConnector/model/TransactionDraft.ts
@@ -1,6 +1,3 @@
-// https://www.npmjs.com/package/@apollo/protobufjs
-import { AccountType } from '@dltConnector/enum/AccountType'
-import { TransactionType } from '@dltConnector/enum/TransactionType'
 import {
   Community as DbCommunity,
   Contribution as DbContribution,
@@ -8,9 +5,10 @@ import {
   User as DbUser,
 } from 'database'
 import { getLogger } from 'log4js'
-import { Duration } from 'shared'
-import { LOG4JS_BASE_CATEGORY_NAME } from '@/config/const'
-import { CODE_VALID_DAYS_DURATION } from '@/graphql/resolver/const/const'
+import { CODE_VALID_DAYS_DURATION, Duration } from 'shared'
+import { LOG4JS_BASE_CATEGORY_NAME } from '../../../config/const'
+import { AccountType } from '../enum/AccountType'
+import { TransactionType } from '../enum/TransactionType'
 import { AccountIdentifier } from './AccountIdentifier'
 import { CommunityAccountIdentifier } from './CommunityAccountIdentifier'
