+++ core/src/apis/federation/client/1_0/SendCoinsClient.ts
@@ -1,9 +1,9 @@
 import { FederatedCommunity as DbFederatedCommunity } from 'database'
 import { GraphQLClient } from 'graphql-request'
 import { getLogger } from 'log4js'
-import { LOG4JS_BASE_CATEGORY_NAME } from '../../../config/const'
-import { EncryptedTransferArgs } from '../../../graphql/model/EncryptedTransferArgs'
-import { ensureUrlEndsWithSlash } from '../../../util/utilities'
+import { LOG4JS_BASE_CATEGORY_NAME } from '../../../../config/const'
+import { EncryptedTransferArgs } from '../../../../graphql/model/EncryptedTransferArgs'
+import { ensureUrlEndsWithSlash } from '../../../../util/utilities'
 import { revertSendCoins as revertSendCoinsQuery } from './query/revertSendCoins'
 import { revertSettledSendCoins as revertSettledSendCoinsQuery } from './query/revertSettledSendCoins'
 import { settleSendCoins as settleSendCoinsQuery } from './query/settleSendCoins'
