+++ core/src/apis/federation/client/1_0/logging/SendCoinsArgsLogging.view.ts
@@ -1,6 +1,6 @@
 import { AbstractLoggingView } from 'database'
 
-import { SendCoinsArgs } from '../model/SendCoinsArgs'
+import { SendCoinsArgs } from '../model'
 
 export class SendCoinsArgsLoggingView extends AbstractLoggingView {
   public constructor(private self: SendCoinsArgs) {
