+++ core/src/apis/federation/client/1_0/model/SendCoinsResult.ts
