+++ core/src/apis/federation/client/1_0/model/index.ts
@@ -1,0 +1,2 @@
+export * from './SendCoinsArgs'
+export * from './SendCoinsResult'
