+++ core/src/apis/federation/client/1_0/query/voteForSendCoins.ts
