+++ core/src/apis/federation/index.ts
@@ -1,0 +1,14 @@
+export { CommandClient as V1_0_CommandClient } from './client/1_0/CommandClient'
+export * from './client/1_0/logging/SendCoinsArgsLogging.view'
+export * from './client/1_0/logging/SendCoinsResultLogging.view'
+export * from './client/1_0/model'
+export * from './client/1_0/query/revertSendCoins'
+export * from './client/1_0/query/revertSettledSendCoins'
+export * from './client/1_0/query/settleSendCoins'
+export * from './client/1_0/query/voteForSendCoins'
+export { SendCoinsClient as V1_0_SendCoinsClient } from './client/1_0/SendCoinsClient'
+export { CommandClient as V1_1_CommandClient } from './client/1_1/CommandClient'
+export { SendCoinsClient as V1_1_SendCoinsClient } from './client/1_1/SendCoinsClient'
+export * from './client/CommandClientFactory'
+export * from './client/SendCoinsClientFactory'
+export * from './enum/apiVersionType'
