+++ core/src/index.ts
@@ -1,24 +1,10 @@
+export * from './apis'
 export * from './command/CommandExecutor'
 export * from './command/CommandFactory'
 export * from './command/commands/SendEmailCommand'
 export * from './command/initCommands'
 export * from './config/index'
 export * from './emails'
-export { CommandClient as V1_0_CommandClient } from './federation/client/1_0/CommandClient'
-export * from './federation/client/1_0/logging/SendCoinsArgsLogging.view'
-export * from './federation/client/1_0/logging/SendCoinsResultLogging.view'
-export * from './federation/client/1_0/model/SendCoinsArgs'
-export * from './federation/client/1_0/model/SendCoinsResult'
-export * from './federation/client/1_0/query/revertSendCoins'
-export * from './federation/client/1_0/query/revertSettledSendCoins'
-export * from './federation/client/1_0/query/settleSendCoins'
-export * from './federation/client/1_0/query/voteForSendCoins'
-export { SendCoinsClient as V1_0_SendCoinsClient } from './federation/client/1_0/SendCoinsClient'
-export { CommandClient as V1_1_CommandClient } from './federation/client/1_1/CommandClient'
-export { SendCoinsClient as V1_1_SendCoinsClient } from './federation/client/1_1/SendCoinsClient'
-export * from './federation/client/CommandClientFactory'
-export * from './federation/client/SendCoinsClientFactory'
-export * from './federation/enum/apiVersionType'
 export * from './graphql/enum/TransactionTypeId'
 export * from './graphql/logic/interpretEncryptedTransferArgs'
 export * from './graphql/logic/processCommand'
@@ -28,8 +14,9 @@ export * from './graphql/logic/storeForeignUser'
 export * from './graphql/model/CommandResult'
 export * from './graphql/model/EncryptedTransferArgs'
 export * from './graphql/scalar/GradidoUnit'
-export * from './logic'
-export * from './schemas'
+export * from './logic/CommunityHandshakeState.logic'
+export * from './logic/community.logic'
+export * from './schemas/transactions.schema'
 export * from './util/calculateSenderBalance'
 export * from './util/utilities'
 export * from './validation/user'
