--- core/src/logic/index.ts
@@ -1,2 +1,0 @@
-export * from './CommunityHandshakeState.logic'
-export * from './community.logic'
